--- a/go.mod
+++ b/go.mod
@@ -1,6 +1,7 @@
 module github.com/quorumcontrol/go-hamt-ipld
 
 require (
+	github.com/gogo/protobuf v1.2.1
 	github.com/ipfs/go-block-format v0.0.2
 	github.com/ipfs/go-blockservice v0.1.1
 	github.com/ipfs/go-cid v0.0.2
@@ -10,6 +11,7 @@ require (
 	github.com/ipfs/go-ipld-cbor v1.5.1-0.20190302174746-59d816225550
 	github.com/ipfs/go-ipld-format v0.0.2
 	github.com/ipfs/go-merkledag v0.1.0
+	github.com/multiformats/go-multihash v0.0.5
 	github.com/polydawn/refmt v0.0.0-20190221155625-df39d6c2d992
 	github.com/spaolacci/murmur3 v1.1.0
 )