… go.sum
… pb/hamt.pb.go