Ok - I think this is a pretty interesting twist... use protobufs underneath the hood for the internal representation of the hamt (not for the values being stored).

Before:
```
$ go test . -run ^TestSetGet -v -count=1
=== RUN   TestSetGet
Total size is: 8340530, size of keys+vals: 6600000, overhead: 1.26
&{4713 100000 map[1:41261 2:16450 3:8613]}
start flush
flush took:  121.143488ms
put of root node start
put of root node took:  1.645556ms  pointer len:  256
finds took:  838.30603ms
--- PASS: TestSetGet (3.68s)
PASS
ok  	github.com/quorumcontrol/go-hamt-ipld	3.702s
```

After:

```
$ go test . -run ^TestSetGet -v -count=1
=== RUN   TestSetGet
Total size is: 8004956, size of keys+vals: 6600000, overhead: 1.21
&{4713 100000 map[1:41261 2:16450 3:8613]}
start flush
flush took:  27.161911ms
put of root node start
put of root node took:  103.124µs  pointer len:  256
finds took:  353.177223ms
--- PASS: TestSetGet (2.05s)
PASS
```