@tobowers 2019-08-18
fwiw: I want to talk this out, not just a quick accept on this. I *think* this makes sense and I don't see a real downside except that it's "different" than the rest of the Dag. And... fwiw: the DagProto format is kind of deprecated at the moment.