@tobowers 2019-08-18
and community for good measure:

Before:
```
goos: darwin
goarch: amd64
pkg: github.com/quorumcontrol/community/hub/tupelocache
BenchmarkAddAddBlockRequest-12    	    5000	   1574492 ns/op	  782081 B/op	   11468 allocs/op
PASS
ok  	github.com/quorumcontrol/community/hub/tupelocache	11.655s
Success: Benchmarks passed.
```


After:
```
oos: darwin
goarch: amd64
pkg: github.com/quorumcontrol/community/hub/tupelocache
BenchmarkAddAddBlockRequest-12    	   10000	    419228 ns/op	  396874 B/op	     547 allocs/op
PASS
ok  	github.com/quorumcontrol/community/hub/tupelocache	11.174s
Success: Benchmarks passed.
```