@tobowers 2019-08-18
If we're going to use a hamt for consensus then we will need to have a much faster version that what's in master now.