@tobowers 2019-08-18
I don’t think these will ever get merged back. The base storage layer is too different already (I didn’t open a PR for those). Moving forward they are going to reimplement a hamt anyway in ipld-prime. They are also staunch cbor supporters