@whyrusleeping 2019-08-20
Just a heads up, i'm going to be redoing the serialization layer for this pretty soon. Early results for other structures being similarly optimized show a 5-10x improvement in marshal speed, and a 3-5x improvement in unmarshal speed.