@whyrusleeping 2019-08-20
Here's one of them: https://github.com/ipfs/go-cid/pull/88

Have some other stuff I havent quite pushed, and this repo is high on my list (its serialization is a bottleneck for us too :grimacing: )